requirements.txt
# Core dependencies
PySide6>=6.5.0
requests>=2.31.0
keyring>=24.2.0
cryptography>=41.0.0
grpcio>=1.57.0
grpcio-tools>=1.57.0
protobuf>=4.24.0

# Database adapters
mysql-connector-python>=8.1.0
oracledb>=1.4.0
pymongo>=4.5.0

# Visualization
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Packaging
PyInstaller>=5.13.0
